Bug 1350858 - Only allow aOpener to be non-null at the first call to SetOpenerWindow(). r?bz
Currently gecko changes the opener of an existing window if it's the target of
a window.open() call or link click. However the spec reads window.opener "must
return the WindowProxy object of the browsing context from which the current
browsing context was created"; not the browsing contexts triggering subsequent
loads.

This patch fixes the behavior by not setting window.opener if it's not the
first load causing window to be created, and not allowing aOpener to be non-null
after first SetOpenerWindow call, with an exception of setting opener from
chrome script.

MozReview-Commit-ID: 6cMmdlSjQ5J