Bug 1314912 - Part 1: Throttle same document navigation caused by content scripts, if there have been too many LOCATION_CHANGE_SAME_DOCUMENT caused by content scripts in a short time frame. r?bz
The patch records the number of successive LOCATION_CHANGE_SAME_DOCUMENT when
the incumbent global exists and it's not system principal, which should
basically indicate content scripts are using History or Location APIs.

When the number of successive LOCATION_CHANGE_SAME_DOCUMENT recorded exceeds
the limit, it throws in InternalLoad and AddState if the caller is from content
scripts and it's making same document navigation.

MozReview-Commit-ID: 7NDp4eOjesp